refactor(config): split config.go (2273 → 226 lines) by concern
config.go had grown to 2273 lines mixing six clearly separable concerns.
Each is now its own file in the same package — no behavior changes,
no exported-symbol renames.

  types.go             types + sentinel errors + default constants
  legacy.go            "device <name> { key = value }" .cfg parser
  yaml_load.go         LoadYAML*, buildConfigFromYAML, base config
  yaml_device.go       per-device YAML conversion + parseDeviceXxx
  yaml_protocols.go    per-protocol YAML→struct converters
  config.go (slim)     Load entry point, lookups, walk validators, utils

go test ./internal/config/... passes; golangci-lint clean.
